refactor(agent): extract issue ranking service
